Rename api to utils folder
